Move strategy to menu
