+++ app/src/main/java/net/chmielowski/randomchoice/ui/screen/input/InputScreen.kt
@@ -249,43 +249,6 @@ internal interface DropdownMenuStrategy {
         @Composable
         override fun Divider(modifier: Modifier) = MenuDefaults.Divider(modifier = modifier)
     }
-
-    // TODO@ Move to test
-    object Fake : DropdownMenuStrategy {
-
-        @Composable
-        override fun Container(content: @Composable () -> Unit) = content()
-
-        @Composable
-        override fun Menu(
-            expanded: Boolean,
-            onDismissRequest: () -> Unit,
-            content: @Composable () -> Unit
-        ) {
-            if (expanded) {
-                content()
-            }
-        }
-
-        @Composable
-        override fun Item(
-            icon: ImageVector,
-            choice: Boolean?,
-            text: Int,
-            onClick: () -> Unit,
-            onDismiss: () -> Unit,
-        ) {
-            Text(
-                stringResource(text),
-                modifier = Modifier.size(10.dp).clickable { onClick();onDismiss() },
-                fontSize = 1.sp
-            )
-        }
-
-        @Composable
-        override fun Divider(modifier: Modifier) {
-        }
-    }
 }
 
 @Composable
